requirements.txt
openai>=1.0.0
tiktoken>=0.5.1
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
PyPDF2>=3.0.0
rich